requirements.txt
feedparser
newspaper3k
lxml_html_clean
openai